requirements.txt
pandas
numpy
beautifulsoup4
lxml
sentence-transformers
transformers
torch
faiss-cpu
pillow
requests
scikit-learn
jupyter
ipykernel
